rename ImageUpload component to InputFileDraggable
