+++ packages/@coorpacademy-components/src/atom/input-file-draggable/style.css
