+++ packages/@coorpacademy-components/src/atom/input-file-draggable/test/fixtures/desktop.js
