+++ packages/@coorpacademy-components/src/atom/input-file-draggable/test/index.js
@@ -3,7 +3,7 @@ import browserEnv from 'browser-env';
 import React from 'react';
 import {render, fireEvent} from '@testing-library/react';
 import {once} from 'lodash/fp';
-import ImageUpload from '..';
+import InputFileDraggable from '..';
 
 browserEnv();
 
@@ -25,7 +25,7 @@ test('should reset Component content', t => {
       t.pass();
     })
   };
-  const {container} = render(<ImageUpload {...props} />);
+  const {container} = render(<InputFileDraggable {...props} />);
   const dragAndDrop = container.querySelector(`[data-name="drag-and-drop-wrapper"]`);
   t.truthy(dragAndDrop);
   const resetIconSearchRes = dragAndDrop.querySelector('[data-name="reset-content-icon"]');
