+++ packages/@coorpacademy-components/src/molecule/drag-and-drop-wrapper/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {map} from 'lodash/fp';
-import ImageUpload from '../../atom/image-upload';
+import InputFileDraggable from '../../atom/input-file-draggable';
 import style from './style.css';
 
 const DragAndDropWrapper = props => {
@@ -10,7 +10,7 @@ const DragAndDropWrapper = props => {
   const dragAndDropList = map(dragAndDropProps => {
     return (
       <li className={style.dragAndDrop} key={`dragAndDrop-${dragAndDropProps.title}`}>
-        <ImageUpload {...dragAndDropProps} />
+        <InputFileDraggable {...dragAndDropProps} />
       </li>
     );
   }, list);
@@ -25,7 +25,7 @@ const DragAndDropWrapper = props => {
 };
 
 DragAndDropWrapper.propTypes = {
-  list: PropTypes.arrayOf(PropTypes.shape(ImageUpload.propTypes)),
+  list: PropTypes.arrayOf(PropTypes.shape(InputFileDraggable.propTypes)),
   'data-name': PropTypes.string
 };
