+++ packages/@coorpacademy-components/src/molecule/drag-and-drop-wrapper/test/fixtures/disabled.js
@@ -1,4 +1,4 @@
-import firstdDragAndDrop from '../../../../atom/image-upload/test/fixtures/default';
+import firstdDragAndDrop from '../../../../atom/input-file-draggable/test/fixtures/default';
 
 export default {
   props: {
