+++ packages/@coorpacademy-components/src/molecule/title-radio-wrapper/types.ts
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import ImageUpload from '../../atom/image-upload';
+import InputFileDraggable from '../../atom/input-file-draggable';
 import RadioWithTitle from '../../atom/radio-with-title';
 import {RadioWithTitleProps} from '../../atom/radio-with-title/types';
 import SelectMultiple from '../select-multiple';
@@ -8,14 +8,14 @@ const propTypes = {
   radioWithTitle: PropTypes.shape(RadioWithTitle.propTypes),
   field: PropTypes.oneOfType([
     PropTypes.shape(SelectMultiple.propTypes),
-    PropTypes.shape(ImageUpload.propTypes)
+    PropTypes.shape(InputFileDraggable.propTypes)
   ]),
   childType: PropTypes.oneOf(['select-multiple', 'massive-upload'])
 };
 
 export type TitleRadioWrapperProps = {
   radioWithTitle: RadioWithTitleProps;
-  field: typeof SelectMultiple.propTypes | typeof ImageUpload.propTypes;
+  field: typeof SelectMultiple.propTypes | typeof InputFileDraggable.propTypes;
   childType: 'select-multiple' | 'massive-upload';
 };
